requirements.txt
streamlit
polars
yfinance
plotly
anthropic
python-dotenv
requests
numpy
orjson